Added a test for, and fixed, a new problem in web.py.
The new logic in web.application._match to enable r"redirect /hello/\1" to be
used as a mapping target (the odd-numbered indices in self.mapping)
unfortunately broke the ability for anything other than a string to be used as
a mapping target; in our case, it broke /blog and everything following it in
the mapping, including /data.  I fixed this so that it doesn't try to do the
substitution unless the target is a string.

I added a test for this in webapp_test.py.  Someone should probably add a test
for it in web.test.application also, so this doesn't get broken again.

I think adding syntax inside the target string is of questionable taste; it
seems to me that it would be much better to write redirect("/hello/") rather
than r"redirect /hello/\1", so that the user isn't tempted to edit
web.application._delegate to add an additional .startswith("redirectpermanent")
test to implement a permanent redirect, or whatever.  (In fact, I think having
strings as mapping targets at all is in questionable taste.)  But I think that's
probably a discussion that needs to happen among the normal web.py developers,
so I'm not taking out this new "redirect .*" feature..