Added support for SQLite 3, with an eye to using it for watchdog unit tests.
SQLite 3 (with pysqlite 1.1.8a-1) supports boolean columns, but it doesn't
support the words True and False to specify values for them.  It treats them as
numeric.  Consequently, when inserting or updating things in the database, we
need to change True and False to 1 and 0, or we get errors from SQLite.

Similarly, if we try to insert Unicode strings, we get errors (this time from
Python).  So this commit adds code to encode the Unicode strings as strings of
bytes before handing them to the database --- by default, as UTF-8.