Got RDF-XML back to working.
The list of interest group rankings had just been getting rendered as a string.
So now the RDF-XML interface produces reasonable RDF-XML for that case,
although not in general for other cases.

I've pasted the N3 and RDF-XML outputs at
http://www.rdfabout.com/demo/validator/ to validate that they parse, but I
haven't written real tests for them yet.  So probably the dumb N3 test that is
here now will break as soon as someone changes something, and it's probably
best to take it out then unless you feel like fixing it.