Some minor fixes in rdftramp.py.
The biggest one was that `super(...).__eq__(...)` only works if your superclass
actually has an `__eq__` method, as opposed to, say, just supporting
comparisons for equality.  In Python 2.4, `unicode` objects in fact don't have
a `__eq__`.  Simpler code: `unicode(self) == other`.

Slightly smaller was the apparently unnecessary code; on my machine, with both
Python 2.4 and Python 2.5, the special case for URI objects was unnecessary, as
indicated by the unit tests I added for that behavior.

Finally, there was a boolean expression in void context, and a comparison of
foo.sort() to bar.sort().  But .sort() returns None, so the comparison doesn't
prove much if it comes out equal.