Some initial work at loading senators.
There are still a few problems: 1. I simply used the state names for the districts, it looks like (from canidates.json) that I should have used the state name with -SEN1 or -SEN2 appended. 2. The /p/by/* pages should seperate out the senate from the house. 3. earmarks, special intrest groups, and ?? aren't done or are not fully functional for the senate.