Got FEC parsing to handle large filings efficiently, for the first time ever.
Previously it choked for several hours (on my machine) on MoveOn's 2004
December 22 filing, 156971.fec, because it was 62 megabytes.  This version is
able to process that filing in less than ten minutes, using about 12 megabytes
of virtual memory for most of that time (although it uses more at the beginning
as it unpacks the zipfile into a temp directory).

However, this is a pretty gross checkin:
- I haven't fixed the tests in fec_crude_csv_test.py, so they crash;
- there's an "assert" to check input validity;
- the command-line interface (if __name__ == '__main__') in fec_crude_csv.py is
  broken (I've been using load/fecdummydb.py);
- the commit gloms together several separate changes, which makes it hard to
  read.

So I'm planning to redo this code change a bit more cleanly, but I'm committing
this so that I have a point of reference in case I get lost in the near future.