Parallelize fec_crude_csv FEC parsing.  On my dual-core Celeron, this cuts about 40% off the run time.
I was going to make it `nice` the worker processes, but then the CPU
clock rate stayed at 1.2GHz instead of jumping up to 1.6GHz.  So now
it just drives the load average up above 8.  Sorry about that.  If I
installed a central work queue that all the workers pulled from, I
could leave one of them non-niced.

It seems to get better than 99% efficiency in CPU usage until only
two processes remain; at that point, it occasionally has moments when
one process is stopped on disk I/O or something, but that's still
better than 95% efficient.  Then of course when only one process
remains, efficiency drops to 50%.

In measurements, there is very little difference on my machine between
running with 2 processes and running with 8.  I'm checking it in with
8 child processes because the penalty on a 2-core machine is
empirically small (maybe 5%) and it will take optimal advantage of
4-core machines, and close to optimal on 8-core machines, and I don't
think any of us have plans to run it on anything bigger than that in
the near future.  Here are my performance measurements:

*0?22.zip: 2m24s real with 8 procs (4m6s user)
3m45s with 1 proc (3m38s user)
2m15s with 2 procs (4m6s user)
2m25s with 3 procs (4m3s user)
2m19s with 4 procs (4m6s user)
2m18s with 5 procs (4m6s user)
2m20s with 6 procs (4m6s user)
2m23s with 7 procs (4m6s user)
2m33s with 8 procs (4m12s user)  (the system wasn't very idle!)
2m29s with 3 procs (4m1s user)
2m16s with 2 procs (4m4s user)
3m5s with 8 procs niced (5m3s user)
2m26s with 8 procs (4m5s user)
2m58s with 8 procs niced again (4m56s)

*??1[38].zip: 20m0s real with 1 proc (19m7s user)
11m38s with 8 procs (21m14s user)